requirements.txt
# Python Development
pre-commit
psutil
pytest
pytest-mpi
tqdm

# Machine Learning
torch==1.13.1
hydra-core
matplotlib
pandas
pickle5
wandb
--extra-index-url https://download.pytorch.org/whl/cu116

# Graph Learning
pyg_lib
torch_scatter==2.1.0+pt113cu116
torch_sparse==0.6.16+pt113cu116
torch_cluster==1.6.0+pt113cu116
torch_spline_conv==1.2.1+pt113cu116
torch_geometric==2.2.0
-f https://data.pyg.org/whl/torch-1.13.0+cu116.html

dgl-cu113
dglgo
-f https://data.dgl.ai/wheels/repo.html

ase
e3nn
geotorch
mendeleev
networkx
ogb